// packages/core/src/enums.ts
export enum VLConnectionCurveEnum {
  BEZIER = 'bezier',
  STRAIGHT = 'straight',
  SMOOTHSTEP = 'smoothstep',
}

export enum VLMarkerShapeEnum {
  CIRCLE = 'circle',
  SQUARE = 'square',
  DIAMOND = 'diamond',
  ARROW = 'arrow',
}

export enum VLFixedSideEnum {
  TOP = 'top',
  RIGHT = 'right',
  BOTTOM = 'bottom',
  LEFT = 'left',
  AUTO = 'auto',
}

export enum VLOrientEnum {
  AUTO = 'auto',
  FIXED = 'fixed',
}